MSVC: re-do getAddrOfVTable() if the vtable global already exists and was considered external to the TU.
getAddrOfVTable has two different behaviors, depending on whether the vtable is considered external or has to be emitted. If getAddrOfVTable() was called before and the VTable has to be emitted, the previous constant needs to be erased and getAddrOfVTable() called again (it will create an internal constant and a global alias taking the vtable name).